+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `paymentAmount` and `paymentCalculation` to `FormSubmissionEventConfiguration`
+- `amountType`, `paymentAmount` and `paymentCalculation` to `FormSubmissionEventConfiguration` for payment event amount configuration
 
 ## [13.1.0] - 2026-07-09