+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `amountType`, `paymentAmount` and `paymentCalculation` to `FormSubmissionEventConfiguration` for payment event amount configuration
+- `screenReaderAlert` to `FormElement`
 
 ## [13.1.0] - 2026-07-09